agent-core/requirements.txt
boto3
strands-agents
strands-agents-tools
bedrock-agentcore
bedrock-agentcore-starter-toolkit
requests
python-dotenv
mcp